Replace TeamTotalRow
